[Chore] conflict 해결

+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/data/dto/remote/response/bookmark/PlaceBookmarkListData.kt
@@ -4,6 +4,6 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class PlaceBookmarkListData(
-    val placeId: Int,
+    val placeId: Long,
     val placeName: String
 )
